pyproject.toml
[tool.poetry]
name = "coda-ai"
version = "0.1.0"
description = ""
authors = ["BuddhiD <buddhid.jaya@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "^0.1.4"
python-dotenv = "^1.0.1"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
redis = "^5.0.1"
openai = "^1.10.0"
langchain-openai = "^0.0.5"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
pymongo = "^4.6.1"
pyjwt = "^2.8.0"
pytz = "^2024.1"
python-dateutil = "^2.8.2"
google-generativeai = "^0.3.2"
langchain-google-genai = "^0.0.9"
dramatiq = "^1.16.0"
chromadb = "^0.4.24"
typer = "^0.9.0"
unstructured = "^0.12.6"
pytube = "^15.0.0"
pandas = "^2.2.1"
sentence-transformers = "^2.5.1"
spacy = "^3.7.4"
langchain-groq = "^0.1.0"
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
